+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 2.0.0-preview2
+## 2.0.0-preview1.1
 ### Additions
 - File Logging to folder `Logs`
   - There is one log file per day
@@ -8,6 +8,12 @@
 - Detection of operating systems in the hardware embed
 - Complete rework of the logic behind `core help` command
 - Code and reliability improvements
+
+### Fixes
+- Database does not save objects when they are modified
+- Debug Servers not showing inside the AutoComplete, affected commands:
+  - `admin debug-servers add-server`
+  - `admin debug-servers remove-server`
 
 ## 2.0.0-preview1
 ### General